Fix Graphics ID expansion

+++ src/field_specials.c
@@ -542,7 +542,7 @@ void SpawnLinkPartnerObjectEvent(void)
     };
     u8 myLinkPlayerNumber;
     u8 playerFacingDirection;
-    u8 linkSpriteId;
+    u16 linkSpriteId;
     u8 i;
 	bool8 foundMatch;
